requirements.txt
autopep8==1.4.3
coverage==4.5.2
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
flake8==3.6.0
mccabe==0.6.1
model-mommy==1.6.0
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.6.4
pytz==2018.7
six==1.11.0
